enhancement/merge-new-develop (#84)
- merge changes made to new-develop into develop
- change github action to trigger from changes to develop branch
- update api url to be set at build time using REACT_APP_API_URL
