+++ src/App.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import { connect } from 'redux-bundler-react';
+
 import Modal from './app-components/modal';
+import NavBar from './app-components/navigation';
+import PageContent from './app-components/pageContent';
 
 import './css/bootstrap/css/bootstrap.water.min.css';
 import './css/mdi/css/materialdesignicons.min.css';
@@ -9,7 +12,10 @@ import './css/index.scss';
 
 export default connect('selectRoute', ({ route: Route }) => (
   <div>
-    <Route />
-    <Modal />
+    <NavBar theme='primary' />
+    <PageContent>
+      <Route />
+    </PageContent>
+    <Modal closeWithEscape />
   </div>
 ));
