+++ src/app-components/card/cardHeader.scss
@@ -1,6 +1,7 @@
 .app-card-header {
   background-color: #f7f7f7;
   border-bottom: 1px solid #c8cacb;
+  display: block;
   margin-bottom: 0;
   padding: 13px 20px;
   z-index: 2;
