+++ src/app-components/modal/index.js
@@ -1,0 +1,3 @@
+export { default } from './modal';
+export { default as ModalHeader } from './modalHeader';
+export { default as ModalFooter } from './modalFooter';
