+++ src/app-components/navigation/devBanner.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './navigation.scss';
+
+const DevBanner = () => (
+  <div className='dev-banner'>
+    <b>DEVELOPMENT ENVIRONMENT</b> - Using only <b>TEST DATA</b>.
+  </div>
+);
+
+export default DevBanner;
