+++ src/app-components/navigation/index.js
@@ -1,0 +1,3 @@
+export { default } from './navBar';
+export { default as NavItem } from './navItem';
+export { default as ProfileMenu} from './profileMenu';
