+++ src/app-pages/collectiongroup/collectiongroup.js
@@ -8,32 +8,30 @@ import collectionGroupForm from '../manager/forms/collection-group-form';
 import collectionGroupTimeseriesPicker from './collectiongroup-timeseries-picker';
 import Icon from '../../app-components/icon';
 import LoginMessage from '../../app-components/login-message';
-import Navbar from '../../app-components/navbar';
 import Notifications from '../../app-components/notifications';
 import RoleFilter from '../../app-components/role-filter';
 import TimeseriesList from './collectiongroup-timeseries-list';
 import TimestampModeSwitcher from './collectiongroup-timestamp-mode-switcher';
 
 export default connect(
-  'doModalOpen',
-  'selectProjectsByRoute',
-  'selectCollectionGroupDetailByRoute',
-  'selectAppTime',
   'doCollectionGroupRemoveTimeseries',
-  'doTimeseriesMeasurementsSave',
+  'doModalOpen',
   'doNotificationFire',
+  'doTimeseriesMeasurementsSave',
+  'selectAppTime',
+  'selectCollectionGroupDetailByRoute',
+  'selectProjectsByRoute',
   ({
-    doModalOpen,
-    projectsByRoute: project,
-    collectionGroupDetailByRoute: detail,
-    appTime,
     doCollectionGroupRemoveTimeseries,
-    doTimeseriesMeasurementsSave,
+    doModalOpen,
     doNotificationFire,
+    doTimeseriesMeasurementsSave,
+    appTime,
+    collectionGroupDetailByRoute: detail,
+    projectsByRoute: project,
   }) => {
     const [isShown, setIsShown] = useState(true);
-    // Valid timestampMode: "now", "choose"
-    const [timestampMode, setTimestampMode] = useState('now');
+    const [timestampMode, setTimestampMode] = useState('now'); // One of ['now', 'choose']
     const [date, setDate] = useState(new Date());
 
     useEffect(
@@ -82,100 +80,93 @@ export default connect(
     };
 
     return (
-      detail && (
-        <>
-          <div style={{ marginBottom: '200px' }}>
-            <Navbar theme='primary' fixed />
-            <section className='container' style={{ marginTop: '6rem' }}>
-              <div className='d-flex flex-row align-items-center'>
-                <h4 className='py-4'>{detail.name}</h4>
-                <RoleFilter
-                  allowRoles={[`${project.slug.toUpperCase()}.*`]}
-                  alt={() => <LoginMessage className='p-2' />}
-                >
-                  <Button
-                    variant='link'
-                    size='small'
-                    className='text-left'
-                    handleClick={(e) => {
-                      doModalOpen(collectionGroupForm, { item: detail });
-                      e.stopPropagation();
+      <div style={{ marginBottom: '200px' }}>
+        {detail && (
+          <section className='container'>
+            <div className='d-flex flex-row align-items-center'>
+              <h4 className='py-4'>{detail.name}</h4>
+              <RoleFilter
+                allowRoles={[`${project.slug.toUpperCase()}.*`]}
+                alt={() => <LoginMessage className='p-2' />}
+              >
+                <Button
+                  variant='link'
+                  size='small'
+                  className='text-left'
+                  handleClick={(e) => {
+                    doModalOpen(collectionGroupForm, { item: detail });
+                    e.stopPropagation();
+                  }}
+                  text='Edit'
+                />
+              </RoleFilter>
+            </div>
+            {/* CONFIGURATION DETAILS */}
+            <div className='row'>
+              <div className='col'>
+                <Card>
+                  <Card.Header
+                    className='bg-light p-2 px-3'
+                    style={{
+                      display: 'flex',
+                      alignItems: 'center',
+                      justifyContent: 'space-between',
+                      cursor: 'pointer',
                     }}
-                    text='Edit'
-                  />
-                </RoleFilter>
-              </div>
-              {/* CONFIGURATION DETAILS */}
-              <div className='row'>
-                <div className='col'>
-                  <Card>
-                    <Card.Header
-                      className='bg-light p-2 px-3'
-                      style={{
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'space-between',
-                        cursor: 'pointer',
-                      }}
-                    >
-                      <div className='d-flex flex-row align-items-center'>
-                        <Icon
-                          icon={isShown ? 'chevron-up' : 'chevron-down'}
-                          className='mdi-24px mr-2'
-                          onClick={(_e) => setIsShown(!isShown)}
+                    onClick={(_e) => setIsShown(!isShown)}
+                  >
+                    <div className='d-flex flex-row align-items-center'>
+                      <Icon
+                        icon={isShown ? 'chevron-up' : 'chevron-down'}
+                        className='mdi-24px mr-2'
+                      />
+                      <h5 className='p-0 m-0'>Timeseries</h5>
+                      <RoleFilter allowRoles={[`${project.slug.toUpperCase()}.*`]}>
+                        <Button
+                          variant='link'
+                          size='small'
+                          handleClick={(e) => {
+                            doModalOpen(collectionGroupTimeseriesPicker);
+                            e.stopPropagation();
+                          }}
+                          text='Add'
+                          icon={<Icon icon='plus' />}
                         />
-                        <div>
-                          <strong className='h5 p-0 m-0'>Timeseries</strong>
-                          <RoleFilter allowRoles={[`${project.slug.toUpperCase()}.*`]}>
-                            <Button
-                              variant='link'
-                              size='small'
-                              handleClick={(e) => {
-                                doModalOpen(collectionGroupTimeseriesPicker);
-                                e.stopPropagation();
-                              }}
-                              text='Add'
-                              icon={<Icon icon='plus' />}
-                            />
-                          </RoleFilter>
-                        </div>
-                      </div>
-                      <div>
-                        <TimestampModeSwitcher
-                          timestampMode={timestampMode}
-                          setTimestampMode={setTimestampMode}
-                          date={date}
-                          setDate={setDate}
-                        />
-                      </div>
-                    </Card.Header>
+                      </RoleFilter>
+                    </div>
+                    <TimestampModeSwitcher
+                      timestampMode={timestampMode}
+                      setTimestampMode={setTimestampMode}
+                      date={date}
+                      setDate={setDate}
+                    />
+                  </Card.Header>
 
-                    {isShown ? (
-                      <Card.Body>
-                        <div style={{ maxHeight: '600px', overflow: 'auto' }}>
-                          <TimeseriesList
-                            items={detail.timeseries}
-                            date={date}
-                            handleItemDelete={(item) => {
-                              doCollectionGroupRemoveTimeseries({
-                                projectId: detail.project_id,
-                                collectionGroupId: detail.id,
-                                timeseriesId: item.id,
-                              });
-                            }}
-                            handleItemSaveValue={handleTimeseriesSaveValue}
-                          />
-                        </div>
-                      </Card.Body>
-                    ) : null}
-                  </Card>
-                </div>
+                  {isShown ? (
+                    <Card.Body>
+                      <div style={{ maxHeight: '600px', overflow: 'auto' }}>
+                        <TimeseriesList
+                          items={detail.timeseries}
+                          date={date}
+                          handleItemDelete={(item) => {
+                            doCollectionGroupRemoveTimeseries({
+                              projectId: detail.project_id,
+                              collectionGroupId: detail.id,
+                              timeseriesId: item.id,
+                            });
+                          }}
+                          handleItemSaveValue={handleTimeseriesSaveValue}
+                        />
+                      </div>
+                    </Card.Body>
+                  ) : null}
+                </Card>
               </div>
-            </section>
-            <Notifications />
-          </div>
-        </>
-      )
+            </div>
+          </section>
+        )}
+        <Notifications />
+      </div>
     );
   }
 );
