+++ src/app-pages/instrument/notes.js
@@ -9,8 +9,6 @@ import RoleFilter from '../../app-components/role-filter';
 const NoteEditor = connect(
   'selectInstrumentsByRoute',
   ({ instrumentsByRoute: instrument, note, onSave, onCancel }) => {
-    if (!instrument) return null;
-
     const [id] = useState(note.id || null);
     const [title, setTitle] = useState(note.title || '');
     const [body, setBody] = useState(note.body || '');
@@ -26,38 +24,40 @@ const NoteEditor = connect(
     };
 
     return (
-      <div style={{ margin: '1em', paddingBottom: '1em' }}>
-        <div className='form-group'>
-          <input
-            className='form-control'
-            type='text'
-            placeholder='Title'
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-          />
+      instrument && (
+        <div style={{ margin: '1em', paddingBottom: '1em' }}>
+          <div className='form-group'>
+            <input
+              className='form-control'
+              type='text'
+              placeholder='Title'
+              value={title}
+              onChange={(e) => setTitle(e.target.value)}
+            />
+          </div>
+          <div className='form-group'>
+            <textarea
+              className='form-control'
+              placeholder='Note...'
+              value={body}
+              onChange={(e) => setBody(e.target.value)}
+              rows={5}
+            />
+          </div>
+          <div>
+            <Button
+              className='mr-2'
+              handleClick={handleSave}
+              text='Save'
+            />
+            <Button
+              variant='secondary'
+              handleClick={onCancel}
+              text='Cancel'
+            />
+          </div>
         </div>
-        <div className='form-group'>
-          <textarea
-            className='form-control'
-            placeholder='Note...'
-            value={body}
-            onChange={(e) => setBody(e.target.value)}
-            rows={5}
-          />
-        </div>
-        <div>
-          <Button
-            className='mr-2'
-            handleClick={handleSave}
-            text='Save'
-          />
-          <Button
-            variant='secondary'
-            handleClick={onCancel}
-            text='Cancel'
-          />
-        </div>
-      </div>
+      )
     );
   }
 );
@@ -65,6 +65,7 @@ const NoteEditor = connect(
 const NoteItem = ({ note, editable, save, del }) => {
   const [isEditing, setEditing] = useState(false);
   const date = new Date(note.time);
+
   return (
     <li className='list-group-item'>
       {isEditing ? (
@@ -116,50 +117,50 @@ export default connect(
     doInstrumentNotesSave: save,
     doInstrumentNotesDelete: del,
   }) => {
-    if (!project) return null;
-
-    const [isAdding, setAdding] = useState(false);
+    const [isAdding, setIsAdding] = useState(false);
 
     const edipi = Number(authEdipi);
     const sorted = notes.sort();
 
     return (
-      <Card>
-        <Card.Header text='Notes' />
-        <ul className='list-group list-group-flush'>
-          {sorted.map((note, i) => (
-            <NoteItem
-              key={i}
-              note={note}
-              save={save}
-              del={del}
-              editable={edipi === note.creator}
-            />
-          ))}
-          <li className='list-group-item'>
-            {isAdding ? (
-              <NoteEditor
-                key={'editor'}
-                note={{}}
-                onSave={(newNote) => {
-                  setAdding(false);
-                  save(newNote);
-                }}
-                onCancel={() => setAdding(false)}
+      project && (
+        <Card>
+          <Card.Header text='Notes' />
+          <ul className='list-group list-group-flush'>
+            {sorted.length ? sorted.map((note, i) => (
+              <NoteItem
+                key={i}
+                note={note}
+                save={save}
+                del={del}
+                editable={edipi === note.creator}
               />
-            ) : (
-              <RoleFilter allowRoles={[`${project.slug.toUpperCase()}.*`]}>
-                <Button
-                  variant='info'
-                  isOutline
-                  text='Add New'
-                  handleClick={() => setAdding(true)}
+            )) : <span className='m-3'>No Notes Added</span>}
+            <li className='list-group-item'>
+              {isAdding ? (
+                <NoteEditor
+                  key={'editor'}
+                  note={{}}
+                  onSave={(newNote) => {
+                    setIsAdding(false);
+                    save(newNote);
+                  }}
+                  onCancel={() => setIsAdding(false)}
                 />
-              </RoleFilter>
-            )}
-          </li>
-        </ul>
-      </Card>
+              ) : (
+                <RoleFilter allowRoles={[`${project.slug.toUpperCase()}.*`]}>
+                  <Button
+                    variant='info'
+                    isOutline
+                    text='Add New'
+                    handleClick={() => setIsAdding(true)}
+                  />
+                </RoleFilter>
+              )}
+            </li>
+          </ul>
+        </Card>
+      )
     );
   }
 );
