+++ src/index.js
@@ -8,14 +8,14 @@ import cache from './cache';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
-cache.getAll().then(initialData => {
+cache.getAll().then((initialData) => {
   const store = getStore(initialData);
 
   if (process.env.NODE_ENV === 'development') window.store = store;
 
   ReactDOM.render(
     <Provider store={store}>
-      <div onClick={getNavHelper(store.doUpdateUrlWithHomepage)}>
+      <div onClick={getNavHelper(store.doUpdateUrl)}>
         <App />
       </div>
     </Provider>,
